Abstract out methods
