+++ src/main/java/duke/tasks/Tentative.java
@@ -5,7 +5,6 @@ import duke.DateTime;
 import java.util.Date;
 
 public class Tentative extends Task {
-    private DateTime endDate;
 
     /**
      * Constructor for Duke.Tasks.Tentative object.
@@ -44,14 +43,6 @@ public class Tentative extends Task {
         return "[TE]" + super.toString() + " (around: "
                 + this.getStartDateString() + " to " + this.getEndDateString()
                 + " - date not fixed)";
-    }
-
-    private String getStartDateString() {
-        return this.startDate.toString();
-    }
-
-    private String getEndDateString() {
-        return this.endDate.toString();
     }
 
     public String getDescription() {
